requirements.txt
aiogram==2.18
psycopg2-binary==2.9.1
PyYAML==6.0
